+++ src/utils/mongo_interface.py
@@ -25,7 +25,7 @@ def getUserfromNick(nickname:str):
     users = client['Comrade']['UserData']
     return users.find_one({"nickname":nickname})
 
-def getCFG(serverID):
+def getCFG(serverID:int):
     '''
     Returns a dictionary with the cfg values for Comrade in a given server
     '''
