+++ src/utils/utilities.py
@@ -1,0 +1,34 @@
+import asyncio
+
+import discord # core to bot
+from discord.ext import commands
+
+'''
+Checks
+'''
+
+# Checks
+
+def isOwner(ctx:discord.Guild):
+    '''
+    Determines whether message author is server owner
+    '''
+    return ctx.author.id == ctx.guild.owner.id
+
+'''
+Message Helpers
+'''
+
+async def delSend(s:str, channel:discord.TextChannel, time:int = 10):
+    '''
+    Sends a message to the desired channel, with a deletion option after a fixed time, default 10 seconds.
+    Standard sending module for Comrade
+    '''
+    msg = await channel.send(s)
+    await asyncio.sleep(time)
+    await msg.add_reaction("🗑️")
+
+async def timedSend(s:str, channel:discord.TextChannel, time:int = 10):
+    '''
+    Sends a message to the desired channel, with deletion after a fixed time, default 10 seconds.
+    '''
